requirements.txt
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
cssselect2==0.7.0
fonttools==4.53.1
html5lib==1.1
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
PyYAML==6.0.1
six==1.16.0
tinycss2==1.3.0
weasyprint==62.1
webencodings==0.5.1
zopfli==0.2.3